Neighbor avoidance improvements
Found several bugs in the original neighbor avoidance code from OpenSteer.

The original comments indicated that in the case of perpendicular paths,
only the slowest vehicle steered behind the fastest. This was actually
wrong, and not only the fastest one was doing it, but if both had
equal speeds then both attempted the same maneuver.  Fixed so that it
behaves as originally described, and in the case where both have the
same speed, a random value is rolled for who will attempt the maneuver.

It also had the problem that if the neighbor was not moving, then its
speed would always be lower than the other vehicle, which would expect
it to get out of the way. Crash.

Finally, neighbor avoidance was steering away in an unit vector, not
accounting for each vehicle's radius.