Added readme, beginning re-engineering
I'm going to be experimenting with separating all the functions
into their own behaviors, as suggested by Laurent Lavigne. It's
an approach that might make it a bit more involved to animate
particles, but allows for a more visually-oriented way to create
the behaviors.

Assume that the code is broken until it's been merged back into
the master branch, as the following commits are likely to be
rough drafts.