Separated movement functions from Vehicle into AutonomousVehicle
Having a separate subclass which actually moves the vehicle will
help us in creating other vehicle interfaces that can be used to
calculate the values needed for interaction with others, while still
keeping the vehicle into player control.  It will also give us
different options as to how to apply the steerings.