Added a Center to vehicles, affected Center and Radius by Scale
A test vehicle model provided by the http://m2h.nl/ guys brought a couple
things to my attention. Namely:

The vehicle assumed that its transform's pivot point was at its center.
On that model's cast, the pivot point was at the base, which meant that
the vehicle's center was at the transform's bottom. This caused some
sphere intersection calculations to calculate against an easy-to-miss
section of the sphere, depending on their radius and speed.

I added a center, much like the one for the SphereCollider, which made me
notice that for consistent behavior the Center and Radius should apply the
transform's scale.  In the case of the Radius, which is a single float,
I'm applying the largest of the scale values.

Added custom editors to assist in setting both.

WIP.

TODO: I still have to update the steering behaviors to use the
ScaledRadius property instead of the Radius one.
TODO: Further test the behaviors.