Rewrote FindNextIntersectionWithSphere
The OpenSteer implementation was misbehaving in some cases, and either
not realizing there was an intersection or realizing it until too
late.   Re-wrote it to be closer to Paul Bourke's derivation,
and taking into account on SphericalObstacleAvoidance the vehicle's
future position.  For some reason the original OpenSteer implementation
didn't, but its design decisions weren't documented.