AutonomousVehicle no longer evaluating HasInertia.
All the steering behaviors I've created assume "Star Wars physics"
for their calculations, where if you stop applying a force to the
object it should stop.  This means that HasInertia should be false.

No longer evaluating the value on AutonomousVehicle to avoid the
confusion caused by the vehicle not acting as expected, and added a
warning on vehicle start.