Increased default steering weight
The default weight of 1 could cause a force that was under the
relevance threshold for some behaviors when we have one steering
force combined with multiple that do not return one.  This will
minimize the odds of that happening.

The ideal solution is to change the force averaging behavior to
disregard the forces returning 0 on the average.  It'll change the
way AutonomousVehicle acts, however, which means I need to add a
value to configure it. Leaving that for later.