Splitting Biped and AutonomousVehicle (WIP, needs refactoring)
My current in-development work had initially been done directly on the
AutonomousVehicle class, but changing how the speed and velocity are
treated would have broken any implementations depending on the previous
approach, like Boids.    I have decided to split them in two classes,
with AutonomousVehicle retaining the old way of doing things and Biped
being the class for vehicles that can alter their velocity vector instead
of just their speed (to, say, side-step).

This is a work in progress, and there is still some code duplication
between both classes which needs to be refactored.