SteerForNeighborGroup optimization: Neighbors are cached on detection
This means that IsInNeighborhood is evaluated when detected, not every
time that the behavior is going to calculate its forces.

This helps in lowering the processing load, but could lead to a case
where a vehicle is beyond the set parameters but still considered a
neighbor.

If this is a concern, make sure that vehicles are detected as often as
the vehicle updates is forces.