MAJOR CHANGE: Vehicles will always use their own mass for calculations
WARNING: Make note if your vehicles' internal mass if you're using it
at all before upgrading to this version.  Hopefully you have these
things under version control.

Previously we used the Vehicle's Rigidbody's mass as the vehicle's own
unless OverrideRigidbodyMass was set to true.  This added the need for
some unnecessary checks, and moves some settings to outside of the
vehicle itself, when it's best to have things consolidated.

Starting on this commit, I am:
- Removing any difference between mass and internal mass.
- Removing OverrideRigidbodyMass.
- Simplifying the Mass getter/setter.

If an agent wants to alter the attached rigidbody's mass, they should
do so directly and explicitly.