New TickedVehicle.Stop method
Previously, when update realized that the vehicle could not move it would
zero its velocity, to avoid having left-over velocity move the vehicle
when it was re-enabled and before the new direction was re-calculated.

I've decided to separate this into an explicit Stop() method, since I
don't really like the idea of having a hidden velocity alteration. Users
may want to move the vehicle somewhere else on their own, while having it
continue at its previous speed when re-enabled.