BREAKING CHANGE: New Min/Max distance parameters for SteerForNeighbors
This allows us to further fine-tune boid behavior, disabling some of the
behaviors in a range (for instance, don't align if the neighbor is too
far, or don't do cohesion if it's too close).

Other advantages:

- Consistent parameter filter naming.
- Previous implementation had two equally named parameters on Separation
  and Cohesion (ComfortDistance) with different semantics.

This will however break your prefabs. Make sure you note down the current
values or are able to diff them from your previous commits.

Note that the Min/Max parameters on SteerForNeighborGroup will still act
as a global filter when detecting neighbors.