Solving an occasional invalid orientation assignment
I was getting some occasional invalid rotation assignments on the
separation examples.  This is likely caused because of a bug I just found
where I was normalizing the OrientationVelocity by the speed, not the
TargetSpeed, which could lead to an invalid orientation velocity if the
speed was 0.

As a backup, I'm also normalizing the orientation velocity before
slerping it.